refactor: 기존 InputManager 클래스명 변경, View에 새로운 InputManager 추가
