+++ src/main/java/vendingmachine/view/InputManager.java
@@ -1,0 +1,12 @@
+package vendingmachine.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputManager {
+    public static String getUserInput(String guide) {
+        OutputManager.printMessage(guide);
+        String userInput = Console.readLine();
+        OutputManager.printBlank();
+        return userInput;
+    }
+}
